Revert the changes from PR #18
This went in with the notion of not breaking Rails apps but it did as
shown in #40. In fact the test in c055d0c clearly gives it away as it
had to include `active_support/time` just for them to pass. That should
have been a red flag. This reverts the changes in that PR so we can
address the ActiveSupport dependency in either a major 3.0 release or
as an optional dependency as part of a new 2.x release.