Update to latest Ruby, JRuby and ActiveSupport versions
Bump jruby 9.1.16.0 to 9.1.17.0

This includes an activesupport5 gemfile for testing against that version. We also exclude that from older versions of Ruby since ActiveSupport 5 doesn't support them.

The commented out line for rbx did not exclude the build itself so that actually causes Travis to run another set of builds against Ruby 1.9.3 which we already specify! No need for this.

Now using ActiveSupport 4.2.x series. We had to bump minitest to v5 as activesupport 4.2 requires it. This explains why builds were installing ActiveSuport 4.0 and not 4.2 despite using the `~>` selector:
```
Bundler could not find compatible versions for gem "minitest":
  In activesupport4.gemfile:
    minitest (~> 4.2)
    activesupport (~> 4.2) was resolved to 4.2.11, which depends on
      minitest (~> 5.1)
```

Specify patch for each minor Ruby version