Fix countless `@name` not initialized warnings
The test suite was displaying tons of warnings in regards to this name accessor. There were 2 different kinds.

The first was the most frequent:
```
/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:42: warning: instance variable @name not initialized
```

And the other was for declaring the getter method override which results in 3 warnings.

```
/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:41: warning: method redefined; discarding old name
276/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:50: warning: method redefined; discarding old name
277/home/travis/build/Rykian/clockwork/test/database_events/test_helpers.rb:64: warning: method redefined; discarding old name
```

While it leads to be a bit more code this seemed the safest and easiest way to make the change with the test helpers and not have to update existing tests and also remove all noisy warnings.